# Make imports dynamic based on command for faster startup