At present `__main__.py` imports all modules each time the `pymobiledevice3` script is invoked. This can be slow on some systems and makes working programmatically with pymobiledevice3 undesirable at scale.

This change makes Click module loading more dynamic and thus improves performance; I'm seeing speed-up of around 1.4 seconds on my test bench (EC2 `t2.xlarge` running AL2 and Python 3.9.16) when averaging all subcommand invocations.

```
pymobiledevice3 activation: Now - 0.62 seconds | Was - 2.05 seconds | Difference - 1.43 seconds
pymobiledevice3 afc: Now - 0.68 seconds | Was - 2.04 seconds | Difference - 1.36 seconds
pymobiledevice3 amfi: Now - 0.61 seconds | Was - 2.05 seconds | Difference - 1.44 seconds
pymobiledevice3 apps: Now - 0.66 seconds | Was - 2.03 seconds | Difference - 1.37 seconds
pymobiledevice3 backup2: Now - 0.66 seconds | Was - 2.05 seconds | Difference - 1.39 seconds
pymobiledevice3 bonjour: Now - 0.62 seconds | Was - 2.06 seconds | Difference - 1.44 seconds
pymobiledevice3 companion: Now - 0.60 seconds | Was - 2.05 seconds | Difference - 1.45 seconds
pymobiledevice3 crash: Now - 0.67 seconds | Was - 2.02 seconds | Difference - 1.35 seconds
pymobiledevice3 developer: Now - 0.92 seconds | Was - 2.02 seconds | Difference - 1.10 seconds
pymobiledevice3 diagnostics: Now - 0.62 seconds | Was - 2.07 seconds | Difference - 1.45 seconds
pymobiledevice3 lockdown: Now - 0.63 seconds | Was - 2.08 seconds | Difference - 1.45 seconds
pymobiledevice3 mounter: Now - 0.71 seconds | Was - 2.05 seconds | Difference - 1.34 seconds
pymobiledevice3 notification: Now - 0.61 seconds | Was - 2.03 seconds | Difference - 1.42 seconds
pymobiledevice3 pcap: Now - 0.62 seconds | Was - 2.09 seconds | Difference - 1.47 seconds
pymobiledevice3 power-assertion: Now - 0.62 seconds | Was - 2.04 seconds | Difference - 1.42 seconds
pymobiledevice3 processes: Now - 0.63 seconds | Was - 2.06 seconds | Difference - 1.43 seconds
pymobiledevice3 profile: Now - 0.62 seconds | Was - 2.04 seconds | Difference - 1.42 seconds
pymobiledevice3 provision: Now - 0.61 seconds | Was - 2.11 seconds | Difference - 1.50 seconds
pymobiledevice3 remote: Now - 1.36 seconds | Was - 2.03 seconds | Difference - .67 seconds
pymobiledevice3 restore: Now - 0.75 seconds | Was - 2.08 seconds | Difference - 1.33 seconds
pymobiledevice3 springboard: Now - 0.62 seconds | Was - 2.05 seconds | Difference - 1.43 seconds
pymobiledevice3 syslog: Now - 0.61 seconds | Was - 2.02 seconds | Difference - 1.41 seconds
pymobiledevice3 usbmux: Now - 0.60 seconds | Was - 2.03 seconds | Difference - 1.43 seconds
pymobiledevice3 webinspector: Now - 1.35 seconds | Was - 2.04 seconds | Difference - .69 seconds
```