@doronz88 2023-11-29
Really liked the idea! Please add a `if __name__ == '__main__'`  so the module can be invoked using `python3 -m pymobiledevice3`