@sbgillett 2023-11-29
> `cli: use dynamic imports to improve performance`

You should be able to [squash and merge](https://docs.github.com/en/pull-requests/collaborating-with-pull-requests/incorporating-changes-from-a-pull-request/about-pull-request-merges#squash-and-merge-your-commits). Alternatively I can cancel this PR and create a new one - let me know.